pyproject.toml
[tool.poetry]
name = "config-validation"
version = "0.1.0"
description = "Demo of loading and validating a config file in python"
authors = ["Matthew Price <matthew.g.price@ons.gov.uk>"]
license = "MIT"
readme = "README.md"
packages = [{include = "config_validation"}]


[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
pydantic = "^1.10.2"
more-itertools = "^9.0.0"
pytest = "^7.2.0"
pydantic-factories = "^1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
